+++ modules/net-vpc/variables.tf
@@ -183,19 +183,8 @@ variable "subnets" {
   default = []
 }
 
-variable "subnets_l7ilb" {
-  description = "List of subnets for private HTTPS load balancer."
-  type = list(object({
-    active        = bool
-    name          = string
-    ip_cidr_range = string
-    region        = string
-  }))
-  default = []
-}
-
-variable "subnets_l7rlb" {
-  description = "List of proxy-only subnets for HTTPS regional load balancers. Note: Only one proxy-only subnet for each VPC network in each region can be active."
+variable "subnets_proxy_only" {
+  description = "List of proxy-only subnets for Regional HTTPS  or Internal HTTPS load balancers. Note: Only one proxy-only subnet for each VPC network in each region can be active."
   type = list(object({
     active        = bool
     name          = string